rag-chat/requirements.txt
qdrant-client==1.9.0
requests==2.32.3
gradio==4.44.1
python-docx
pypdf
python-dotenv
openai
huggingface-hub==0.24.0
sentence-transformers==2.7.0
